Extract IDrawable interface from BrushStroke
This provides a common abstraction that can be used for other drawing tools such as gradient.

For #29

+++ Fotografix.Editor/Tools/IBrushToolSettings.cs
@@ -4,7 +4,7 @@ namespace Fotografix.Editor.Tools
 {
     public interface IBrushToolSettings
     {
-        float Size { get; set; }
+        int Size { get; set; }
         Color Color { get; set; }
     }
 }
